+++ Src/Views/TInverseView.cpp
@@ -2360,7 +2360,7 @@ if ( LButtonDown || MButtonDown || RButtonDown ) {
             BFont->SetBoxColor ( GLBASE_CURSORHINTBACKCOLOR );
 
             glTranslated ( MousePos.X(), PaintRect.Height() - MousePos.Y(), 0 );
-            glScaled ( CursorHintSize, CursorHintSize, 1 );
+            glScaled     ( CursorHintSize, CursorHintSize, 1 );
 
             BFont->Print ( 0, -1, 1, "No coordinates available in slice mode.", TA_CENTER | TA_TOP | TA_BOX );
 
@@ -2836,8 +2836,8 @@ if ( RButtonDown ) {
 
         if ( MouseAxis == MouseAxisHorizontal ) {
 
-            if ( dx > 0 )   ScalingLevel /= 1 + (double) min ( (double) adx, MouseMoveScale ) / MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
-            else            ScalingLevel *= 1 + (double) min ( (double) adx, MouseMoveScale ) / MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
+            if ( dx > 0 )   ScalingLevel /= 1 + min ( adx, MouseMoveScale ) / (double) MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
+            else            ScalingLevel *= 1 + min ( adx, MouseMoveScale ) / (double) MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
 
             SetScaling ( ScalingLevel );
             }
